requirements.txt
# Breaking B.A.D. Backend Dependencies
# Pinned versions for reproducible builds

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0

# Request handling
python-multipart==0.0.20

# Database
pymongo==4.10.1

# AI/ML
google-genai==0.5.0

# PDF processing
pypdf==5.1.0

# Configuration
python-dotenv==1.0.1
pydantic-settings==2.7.1
